refactor: replace wrap/wrapETH aliases with standalone shield/shieldETH hooks
Move full implementation into use-shield.ts and use-shield-eth.ts instead of
re-exporting from use-wrap files. Delete use-wrap.ts and use-wrap-eth.ts,
update all imports and tests to use the shield naming directly.
